:hammer: minor progress on live-map
